Fix #27 validation rules are not deferred
